+++ modules/services/model/src/test/java/au/com/shiftyjelly/pocketcasts/models/to/ChapterTest.kt
@@ -7,7 +7,7 @@ import org.junit.Test
 
 class ChapterTest {
     @Test
-    fun calculateProgress() {
+    fun `calculate progress`() {
         val chapter = Chapter(
             title = "",
             startTime = 5.seconds,
@@ -22,18 +22,37 @@ class ChapterTest {
     }
 
     @Test
-    fun remainingTime() {
+    fun `calculate remainging time`() {
         val chapter = Chapter(
             title = "",
             startTime = 5.seconds,
             endTime = 70.seconds,
         )
 
-        assertEquals("1m", chapter.remainingTime(5000.milliseconds))
-        assertEquals("1m", chapter.remainingTime(10000.milliseconds))
-        assertEquals("59s", chapter.remainingTime(11000.milliseconds))
-        assertEquals("2s", chapter.remainingTime(68000.milliseconds))
-        assertEquals("1s", chapter.remainingTime(69000.milliseconds))
-        assertEquals("0s", chapter.remainingTime(69999.milliseconds))
+        assertEquals("1m", chapter.remainingTime(5000.milliseconds, playbackSpeed = 1.0))
+        assertEquals("1m", chapter.remainingTime(10000.milliseconds, playbackSpeed = 1.0))
+        assertEquals("59s", chapter.remainingTime(11000.milliseconds, playbackSpeed = 1.0))
+        assertEquals("2s", chapter.remainingTime(68000.milliseconds, playbackSpeed = 1.0))
+        assertEquals("1s", chapter.remainingTime(69000.milliseconds, playbackSpeed = 1.0))
+        assertEquals("0s", chapter.remainingTime(69999.milliseconds, playbackSpeed = 1.0))
+    }
+
+    @Test
+    fun `calclucate remainging time using playback speed`() {
+        val chapter = Chapter(
+            title = "",
+            startTime = 0.seconds,
+            endTime = 120.seconds,
+        )
+
+        assertEquals("4m", chapter.remainingTime(0.seconds, playbackSpeed = 0.5))
+        assertEquals("2m", chapter.remainingTime(0.seconds, playbackSpeed = 1.0))
+        assertEquals("1m", chapter.remainingTime(0.seconds, playbackSpeed = 1.5))
+        assertEquals("1m", chapter.remainingTime(0.seconds, playbackSpeed = 2.0))
+
+        assertEquals("2m", chapter.remainingTime(75.seconds, playbackSpeed = 0.5))
+        assertEquals("45s", chapter.remainingTime(75.seconds, playbackSpeed = 1.0))
+        assertEquals("30s", chapter.remainingTime(75.seconds, playbackSpeed = 1.5))
+        assertEquals("23s", chapter.remainingTime(75.seconds, playbackSpeed = 2.0))
     }
 }